Extract testAbandonedQueries to a separate test class
This test requires a specific client timeout that interferes with other tests
like `TestServer.testVersionOnCompilerFailedError` that can fail with an abandoned client timeout:

```
Expecting actual:
  "io.trino.spi.TrinoException: Query 20260318_190211_00017_fxacu was abandoned by the client, as it may have exited or stopped checking for query results. Query results have not been accessed since 2026-03-18T19:02:12.074039520Z: currentTime 2026-03-18T19:02:13.575091194Z
	at io.trino.execution.QueryTracker.failAbandonedQueries(QueryTracker.java:276)
	at io.trino.execution.QueryTracker.lambda$start$0(QueryTracker.java:83)
	at com.google.common.util.concurrent.MoreExecutors$ScheduledListeningDecorator$NeverSuccessfulListenableFutureTask.run(MoreExecutors.java:639)
	at java.base/java.util.concurrent.Executors$RunnableAdapter.call(Executors.java:545)
	at java.base/java.util.concurrent.FutureTask.runAndReset(FutureTask.java:369)
	at java.base/java.util.concurrent.ScheduledThreadPoolExecutor$ScheduledFutureTask.run(ScheduledThreadPoolExecutor.java:310)
	at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1090)
	at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:614)
	at java.base/java.lang.Thread.run(Thread.java:1474)
"
to contain pattern:
  "TrinoException: Failed to execute query; (?s:.*)at io.trino.sql.gen.PageFunctionCompiler.compileProjectionInternal"
```
